Add first ISO Inject implementation, partially working.
- Right now it only works with Ubuntu Desktop ISO.
The experiment I wanted to do is to modify an Ubuntu Live Image by
adding some StarlingX deb Packages. Then boot the image and install
it.
What I see is that after the installation the disk is capable of
booting with the Stx Packages installed.

- It is not working with an Ubuntu Server ISO.
I guess it is because for server the iso doesn't have a Try Live
Image option, therefore I'm corrupting the ISO when injecting the
Stx Packages in the wrong directory.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>