Adding Ubuntu Server 18.04 support
Motive: I read somewhere on the Internet that starting on
Ubuntu Server 18.04, the iso has support for Live CD usage.

The experiment in this commit was to use that live server ISO as
the base, then inject the StarlingX packages into its rootfs. Same
as with the Ubuntu Desktop ISO (the one that works).

This change is able to produce an installable ISO with the StarlingX
fmMgr and fm-common installed.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>